task1/requirements.txt
streamlit
requests
beautifulsoup4
transformers
huggingface_hub
lxml
python-dotenv